+++ models/CalendarPlusEventsModel.php
@@ -21,27 +21,25 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniqueCityNamesByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
         $t    = static::$strTable;
-        $time = time();
+        $intStart = time();
+        $intEnd   = 99999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.city != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'city';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         return static::findBy($arrColumns, null, $arrOptions);
@@ -49,39 +47,36 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniquePromotersByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
         $t    = static::$strTable;
-        $time = time();
+        $intStart = time();
+        $intEnd   = 999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.promoter IS NOT NULL)";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return $objEvents;
         }
 
         $arrPromoters = [];
 
-        while ($objEvents->next())
-        {
-            $arrPromoters = array_merge($arrPromoters, deserialize($objEvents->promoter, true));
+        while ($objEvents->next()) {
+            $objEvent = $objEvents->current();
+            $arrPromoters = array_merge($arrPromoters, deserialize($objEvent->promoter, true));
         }
 
         $arrPromoters = array_unique($arrPromoters);
@@ -91,40 +86,36 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniqueDocentsByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
-        $t    = static::$strTable;
-        $time = time();
+        $t        = static::$strTable;
+        $intStart = time();
+        $intEnd   = 99999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.docents != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'docents';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return null;
         }
 
         $arrDocents = [];
 
-        while ($objEvents->next())
-        {
+        while ($objEvents->next()) {
             $arrDocents = array_merge($arrDocents, deserialize($objEvents->docents, true));
         }
 
@@ -135,41 +126,40 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniqueMemberDocentsByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
-        $t    = static::$strTable;
-        $time = time();
+        $t        = static::$strTable;
+        $intStart = time();
+        $intEnd   = 999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.memberDocents != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'memberDocents';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return null;
         }
 
         $arrDocents = [];
 
-        while ($objEvents->next())
-        {
-            $arrDocents = array_merge($arrDocents, deserialize($objEvents->memberDocents, true));
+        $objEvents->reset();
+
+        while ($objEvents->next()) {
+            $objEvent = $objEvents->current();
+            $arrDocents = array_merge($arrDocents, deserialize($objEvent->memberDocents, true));
         }
 
         $arrDocents = array_unique($arrDocents);
@@ -179,41 +169,38 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniqueHostsByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
         $t    = static::$strTable;
-        $time = time();
+        $intStart = time();
+        $intEnd   = 999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.hosts != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'hosts';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return null;
         }
 
         $arrHosts = [];
 
-        while ($objEvents->next())
-        {
-            $arrHosts = array_merge($arrHosts, deserialize($objEvents->hosts, true));
+        while ($objEvents->next()) {
+            $objEvent = $objEvents->current();
+            $arrHosts = array_merge($arrHosts, deserialize($objEvent->hosts, true));
         }
 
         $arrHosts = array_unique($arrHosts);
@@ -223,41 +210,38 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniqueMemberHostsByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
         $t    = static::$strTable;
-        $time = time();
+        $intStart = time();
+        $intEnd   = 999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.memberHosts != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'memberHosts';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return null;
         }
 
         $arrHosts = [];
 
-        while ($objEvents->next())
-        {
-            $arrHosts = array_merge($arrHosts, deserialize($objEvents->memberHosts, true));
+        while ($objEvents->next()) {
+            $objEvent = $objEvents->current();
+            $arrHosts = array_merge($arrHosts, deserialize($objEvent->memberHosts, true));
         }
 
         $arrHosts = array_unique($arrHosts);
@@ -268,41 +252,38 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getUniquePromoterNamesByPids(array $arrPids = [], $currentOnly = true, $arrOptions = [])
     {
-        if (!is_array($arrPids) || empty($arrPids))
-        {
+        if (!is_array($arrPids) || empty($arrPids)) {
             return null;
         }
 
         $t    = static::$strTable;
-        $time = time();
+        $intStart = time();
+        $intEnd   = 999999999999;
 
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumns[] = "($t.promoter != '')";
 
-        if ($currentOnly)
-        {
-            $arrColumns[] = "($t.endTime>=$time OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$time)))";
+        if ($currentOnly) {
+            $arrColumns[] = "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring='1' AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
         }
 
         $arrOptions['group'] = 'promoter';
 
-        if (!BE_USER_LOGGED_IN)
-        {
-            $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
+        if (!BE_USER_LOGGED_IN) {
+            $arrColumns[] = "($t.start='' OR $t.start<$intStart) AND ($t.stop='' OR $t.stop>$intStart) AND $t.published=1";
         }
 
         $objEvents = static::findBy($arrColumns, null, $arrOptions);
 
-        if ($objEvents === null)
-        {
+        if ($objEvents === null) {
             return $objEvents;
         }
 
         $arrPromoters = [];
 
-        while ($objEvents->next())
-        {
-            $arrPromoters = array_merge($arrPromoters, deserialize($objEvents->promoter, true));
+        while ($objEvents->next()) {
+            $objEvent = $objEvents->current();
+            $arrPromoters = array_merge($arrPromoters, deserialize($objEvent->promoter, true));
         }
 
         $arrPromoters = array_unique($arrPromoters);
@@ -341,7 +322,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
      * @param array $arrFilterOptions
      * @param array $arrFilterConfig
      * @param array $arrOptions
-     * @param bool  $count
+     * @param bool $count
      *
      * @return \CalendarEventsModel|\CalendarEventsModel[]|Collection|null|int|static
      */
@@ -373,41 +354,32 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $arrColumns[] = "($t.pid IN (" . implode(',', $arrPids) . "))";
         $arrColumnsOr = [];
 
-        foreach ($arrFilter as $key => $value)
-        {
+        foreach ($arrFilter as $key => $value) {
 
             $arrValueOptions = (isset($arrFilterOptions[$key]) && is_array($arrFilterOptions[$key])) ? $arrFilterOptions[$key] : [];
 
-            switch ($key)
-            {
+            switch ($key) {
                 case 'startDate':
-                    if ($value && $value >= $intStart)
-                    {
+                    if ($value && $value >= $intStart) {
                         $intStart = $value;
                     }
                     break;
                 case 'endDate':
-                    if ($value && $value <= $intEnd)
-                    {
+                    if ($value && $value <= $intEnd) {
                         $intEnd = strtotime(date('d.m.Y', $value) . ' 23:59:59'); // until last second of the day
                     }
                     break;
 
                 case 'dates':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $strQuery   = '';
                         $valueArray = trimsplit(',', $value);
 
-                        foreach ($valueArray as $key => $strDate)
-                        {
+                        foreach ($valueArray as $key => $strDate) {
                             $intDate = strtotime($strDate);
-                            if ($key > 0)
-                            {
+                            if ($key > 0) {
                                 $strQuery .= " OR ((($t.endDate IS NULL OR $t.endDate = '') AND $t.startDate = $intDate) OR (($t.startDate = '' OR $t.startDate <= $intDate) AND $t.endDate >= $intDate))";
-                            }
-                            else
-                            {
+                            } else {
                                 $strQuery =
                                     "((($t.endDate IS NULL OR $t.endDate = '') AND $t.startDate = $intDate) OR (($t.startDate = '' OR $t.startDate <= $intDate) AND $t.endDate >= $intDate))";
                             }
@@ -418,139 +390,99 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
                 case 'promoter':
                 case 'areasoflaw':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $valueArray = trimsplit(',', $value);
 
                         // permit intersections only
-                        if (!empty($arrValueOptions))
-                        {
+                        if (!empty($arrValueOptions)) {
                             $valueArray = array_intersect($valueArray, $arrValueOptions);
                         }
 
-                        if (is_array($valueArray) && !empty($valueArray))
-                        {
-                            if ($arrFilterConfig['show_related'])
-                            {
+                        if (is_array($valueArray) && !empty($valueArray)) {
+                            if ($arrFilterConfig['show_related']) {
                                 $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.$key", $valueArray, EVENTMODEL_CONDITION_OR);
-                            }
-                            else
-                            {
+                            } else {
                                 $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.$key", $valueArray);
                             }
 
                         }
-                    }
-                    else
-                    {
-                        if (!empty($arrValueOptions))
-                        {
+                    } else {
+                        if (!empty($arrValueOptions)) {
                             $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.$key", $arrValueOptions);
                         }
                     }
                     break;
                 case strrpos($key, 'eventtypes', -strlen($key)) !== false :
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $valueArray = trimsplit(',', $value);
                         // permit intersections only
-                        if (!empty($arrValueOptions))
-                        {
+                        if (!empty($arrValueOptions)) {
                             $valueArray = array_intersect($valueArray, $arrValueOptions);
                         }
 
-                        if (is_array($valueArray) && !empty($valueArray))
-                        {
-                            if ($arrFilterConfig['show_related'])
-                            {
+                        if (is_array($valueArray) && !empty($valueArray)) {
+                            if ($arrFilterConfig['show_related']) {
                                 $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.eventtypes", $valueArray, EVENTMODEL_CONDITION_OR);
-                            }
-                            else
-                            {
+                            } else {
                                 $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.eventtypes", $valueArray, EVENTMODEL_CONDITION_AND);
                             }
 
                         }
-                    }
-                    else
-                    {
-                        if (!empty($arrValueOptions))
-                        {
+                    } else {
+                        if (!empty($arrValueOptions)) {
                             $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.eventtypes", $arrValueOptions);
                         }
                     }
                     break;
                 case 'docents':
                 case 'hosts':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $valueArray = trimsplit(',', $value);
 
-                        if (is_array($valueArray) && !empty($valueArray))
-                        {
+                        if (is_array($valueArray) && !empty($valueArray)) {
                             $arrDocents = [];
                             $arrMembers = [];
                             $arrHosts   = [];
 
-                            foreach ($valueArray as $id)
-                            {
+                            foreach ($valueArray as $id) {
                                 // docent
-                                if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'd'))
-                                {
+                                if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'd')) {
                                     $arrDocents[] = substr($id, 1);
-                                }
-                                // members
-                                else
-                                {
-                                    if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'm'))
-                                    {
+                                } // members
+                                else {
+                                    if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'm')) {
                                         $arrMembers[] = substr($id, 1);
-                                    }
-                                    // hosts
-                                    else
-                                    {
-                                        if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'h'))
-                                        {
+                                    } // hosts
+                                    else {
+                                        if (\HeimrichHannot\Haste\Util\StringUtil::startsWith($id, 'h')) {
                                             $arrHosts[] = substr($id, 1);
                                         }
                                     }
                                 }
                             }
 
-                            if (!empty($arrDocents))
-                            {
-                                if (is_array($valueArray) && !empty($valueArray))
-                                {
-                                    if ($arrFilterConfig['show_related'])
-                                    {
+                            if (!empty($arrDocents)) {
+                                if (is_array($valueArray) && !empty($valueArray)) {
+                                    if ($arrFilterConfig['show_related']) {
                                         $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.docents", $arrDocents, EVENTMODEL_CONDITION_OR);
-                                    }
-                                    else
-                                    {
+                                    } else {
                                         $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.docents", $arrDocents);
                                     }
                                 }
                             }
 
-                            if (!empty($arrMembers))
-                            {
-                                if (is_array($valueArray) && !empty($valueArray))
-                                {
-                                    if ($arrFilterConfig['show_related'] || $arrFilterConfig['combine_docents'])
-                                    {
+                            if (!empty($arrMembers)) {
+                                if (is_array($valueArray) && !empty($valueArray)) {
+                                    if ($arrFilterConfig['show_related'] || $arrFilterConfig['combine_docents']) {
                                         $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.memberdocents", $arrMembers, EVENTMODEL_CONDITION_OR);
 
-                                        if ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['combine_docents'])
-                                        {
+                                        if ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['combine_docents']) {
                                             $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.memberhosts", $arrMembers, EVENTMODEL_CONDITION_OR);
                                         }
-                                    }
-                                    else
-                                    {
+                                    } else {
                                         $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.memberdocents", $arrMembers);
 
-                                        if ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['combine_docents'])
-                                        {
+                                        if ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['combine_docents']) {
                                             $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.memberhosts", $arrMembers);
                                         }
 
@@ -558,16 +490,11 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
                                 }
                             }
 
-                            if (!empty($arrHosts) && ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['cal_docent_combine']))
-                            {
-                                if (is_array($valueArray) && !empty($valueArray))
-                                {
-                                    if ($arrFilterConfig['show_related'])
-                                    {
+                            if (!empty($arrHosts) && ($key = 'hosts' || $key == 'docents' && $arrFilterConfig['cal_docent_combine'])) {
+                                if (is_array($valueArray) && !empty($valueArray)) {
+                                    if ($arrFilterConfig['show_related']) {
                                         $arrColumnsOr[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.hosts", $arrHosts, EVENTMODEL_CONDITION_OR);
-                                    }
-                                    else
-                                    {
+                                    } else {
                                         $arrColumns[] = EventModelHelper::createMySQLRegexpForMultipleIds("$t.hosts", $arrHosts);
                                     }
                                 }
@@ -576,24 +503,20 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
                     }
                     break;
                 case 'postal':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $arrColumns[] = "LEFT($t.postal, 1) = ?";
                         $arrValues[]  = $value;
                     }
                     break;
                 case 'timeHours':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $arrColumns[] = "($t.$key >= $value AND $t.$key < ($value + 1))";
                     }
                     break;
                 case 'q':
-                    if ($value != '')
-                    {
+                    if ($value != '') {
 
-                        try
-                        {
+                        try {
                             global $objPage;
 
                             $objSearch = \Search::searchFor(
@@ -606,11 +529,10 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
                             );
 
                             // title / teaser search as fallback
-                            if ($objSearch->numRows < 1)
-                            {
+                            if ($objSearch->numRows < 1) {
                                 $arrColumns[] = "$t.title LIKE ? OR $t.teaser LIKE ?";
-                                $arrValues[] = "%". $value . "%";
-                                $arrValues[] = "%" . $value . "%";
+                                $arrValues[]  = "%" . $value . "%";
+                                $arrValues[]  = "%" . $value . "%";
                                 break;
                             }
 
@@ -620,8 +542,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
                             $n = 0;
 
-                            foreach ($arrUrls as $i => $strAlias)
-                            {
+                            foreach ($arrUrls as $i => $strAlias) {
                                 $strKeyWordColumns .= ($n > 0 ? " OR " : "") . "$t.alias = ?";
                                 $arrParts          = parse_url($strAlias);
                                 $arrValues[]       = basename($arrParts['path']);
@@ -630,15 +551,13 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
                             $arrColumns[] = "($strKeyWordColumns)";
 
-                        } catch (\Exception $e)
-                        {
+                        } catch (\Exception $e) {
                             \System::log('Website search failed: ' . $e->getMessage(), __METHOD__, TL_ERROR);
                         }
                     }
                     break;
                 default:
-                    if ($value != '')
-                    {
+                    if ($value != '') {
                         $arrColumns[] = "$t.$key=?";
                         $arrValues[]  = $value;
                     }
@@ -648,27 +567,23 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $arrColumns[] =
             "(($t.startTime>=$intStart AND $t.startTime<=$intEnd) OR ($t.endTime>=$intStart AND $t.endTime<=$intEnd) OR ($t.startTime<=$intStart AND $t.endTime>=$intEnd) OR ($t.recurring=1 AND ($t.recurrences=0 OR $t.repeatEnd>=$intStart) AND $t.startTime<=$intEnd))";
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = time();
             $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
         }
 
         // for related search
-        if (!empty($arrColumnsOr))
-        {
+        if (!empty($arrColumnsOr)) {
             $arrColumns[] = implode(' OR ', $arrColumnsOr);
         }
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "listTime ASC";
 
 //            $arrOptions['order'] = "$t.startTime";
         }
 
-        if ($count)
-        {
+        if ($count) {
             return static::countBy($arrColumns, $arrValues, $arrOptions);
         }
 
@@ -676,15 +591,13 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
             sprintf("SELECT %s FROM $t WHERE %s ORDER BY %s", implode(',', $arrFields), implode(' AND ', $arrColumns), $arrOptions['order'])
         );
 
-        if ($arrOptions['limit'] > 0)
-        {
+        if ($arrOptions['limit'] > 0) {
             $objStatement->limit($arrOptions['limit'], $arrOptions['offset'] ?: 0);
         }
 
         $objResult = $objStatement->execute($arrValues);
 
-        if ($objResult !== null)
-        {
+        if ($objResult !== null) {
             return Collection::createFromDbResult($objResult, $t);
         }
 
@@ -694,8 +607,8 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     /**
      * Find published events by their parent ID
      *
-     * @param integer $intPid     The calendar ID
-     * @param array   $arrOptions An optional options array
+     * @param integer $intPid The calendar ID
+     * @param array $arrOptions An optional options array
      *
      * @return \CalendarEventsModel|\Model\Collection|null A collection of models or null if there are no events
      */
@@ -704,14 +617,12 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $t          = static::$strTable;
         $arrColumns = ["$t.pid=?"];
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = \Date::floorToMinute();
             $arrColumns[] = "($t.start='' OR $t.start<='$time') AND ($t.stop='' OR $t.stop>'" . ($time + 60) . "') AND $t.published='1'";
         }
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "$t.startTime DESC";
         }
 
@@ -721,7 +632,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     /**
      * Find published events by id or alias
      *
-     * @param mixed $varId      The numeric ID or alias name
+     * @param mixed $varId The numeric ID or alias name
      * @param array $arrOptions An optional options array
      *
      * @return \Model|null The model or null if there is no event
@@ -731,8 +642,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $t          = static::$strTable;
         $arrColumns = ["($t.id=?) OR ($t.alias=?)"];
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = time();
             $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
         }
@@ -742,15 +652,11 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function hasAtLeastOnePublishedPlacedSubEvent($config, $intId)
     {
-        if (in_array('event_subscription', $config->getActiveModules()))
-        {
+        if (in_array('event_subscription', $config->getActiveModules())) {
             $objSubEvents = static::findPublishedSubEvents($intId);
-            if ($objSubEvents !== null)
-            {
-                while ($objSubEvents->next())
-                {
-                    if ($objSubEvents->addSubscription)
-                    {
+            if ($objSubEvents !== null) {
+                while ($objSubEvents->next()) {
+                    if ($objSubEvents->addSubscription) {
                         return true;
                     }
                 }
@@ -763,7 +669,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     /**
      * Find published sub events by the parent's ID or alias
      *
-     * @param mixed $varId      The numeric ID or alias name
+     * @param mixed $varId The numeric ID or alias name
      * @param array $arrOptions An optional options array
      *
      * @return \Model|null The model or null if there is no event
@@ -773,13 +679,11 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $t          = static::$strTable;
         $arrColumns = ["($t.parentEvent=?)"];
 
-        if (!isset($arrOptions['order']))
-        {
+        if (!isset($arrOptions['order'])) {
             $arrOptions['order'] = "$t.startTime";
         }
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = time();
             $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
         }
@@ -790,8 +694,8 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     /**
      * Find published sub events by the parent's ID or alias
      *
-     * @param mixed $varId      The numeric ID or alias name
-     * @param array $arrPids    An array of calendar IDs
+     * @param mixed $varId The numeric ID or alias name
+     * @param array $arrPids An array of calendar IDs
      * @param array $arrOptions An optional options array
      *
      * @return \Model|null The model or null if there is no event
@@ -801,14 +705,10 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $t              = static::$strTable;
         $objEvents      = static::findAll();
         $parentalEvents = [];
-        if ($objEvents !== null)
-        {
-            while ($objEvents->next())
-            {
-                if ($objEvents->parentEvent)
-                {
-                    if (!in_array($objEvents->parentEvent, $parentalEvents))
-                    {
+        if ($objEvents !== null) {
+            while ($objEvents->next()) {
+                if ($objEvents->parentEvent) {
+                    if (!in_array($objEvents->parentEvent, $parentalEvents)) {
                         $parentalEvents[] = $objEvents->parentEvent;
                     }
                 }
@@ -816,8 +716,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         }
         $arrColumns = ["($t.id IN (" . implode(',', $parentalEvents) . "))"];
 
-        if (!BE_USER_LOGGED_IN)
-        {
+        if (!BE_USER_LOGGED_IN) {
             $time         = time();
             $arrColumns[] = "($t.start='' OR $t.start<$time) AND ($t.stop='' OR $t.stop>$time) AND $t.published=1";
         }
@@ -828,10 +727,8 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     public static function hasTextOrTeaser($intId)
     {
         $objEvent = static::findByPk($intId);
-        if ($objEvent !== null)
-        {
-            if ($objEvent->teaser)
-            {
+        if ($objEvent !== null) {
+            if ($objEvent->teaser) {
                 return true;
             }
         }
@@ -844,8 +741,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     public static function getPlacesLeft($intId, $database)
     {
         $objEvent = static::findByPk($intId);
-        if ($objEvent !== null)
-        {
+        if ($objEvent !== null) {
             return $objEvent->placesTotal - static::getReservedPlaces($intId, $database);
         }
 
@@ -856,21 +752,15 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     {
         $objEvent       = static::findByPk($intId);
         $reservedPlaces = 0;
-        if ($objEvent !== null)
-        {
-            if ($useMemberGroups)
-            {
+        if ($objEvent !== null) {
+            if ($useMemberGroups) {
                 $objMembers = $database->prepare('SELECT * FROM tl_member WHERE groups LIKE ?')->execute('%"' . $objEvent->memberGroup . '"%');
 
                 return $objMembers->numRows;
-            }
-            else
-            {
+            } else {
                 $objSubscriber = $database->prepare('SELECT * FROM tl_formdata_details WHERE ff_name=? AND value=?')->executeUncached('eventAlias', $objEvent->alias);
-                if ($objSubscriber->numRows > 0)
-                {
-                    while ($objSubscriber->next())
-                    {
+                if ($objSubscriber->numRows > 0) {
+                    while ($objSubscriber->next()) {
                         $reservedPlaces += 1;
                     }
                 }
@@ -884,8 +774,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     {
         $objEvent = static::findByPk($intId);
 
-        if ($objEvent !== null)
-        {
+        if ($objEvent !== null) {
             $objMembers = \Database::getInstance()->prepare('SELECT * FROM tl_member WHERE groups LIKE ?')->execute('%"' . $objEvent->memberGroupCheckedIn . '"%');
 
             return $objMembers->numRows;
@@ -897,8 +786,7 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     public static function getPlacesLeftSubEvent($intId, $database)
     {
         $objEvent = static::findByPk($intId);
-        if ($objEvent !== null)
-        {
+        if ($objEvent !== null) {
             return $objEvent->placesTotal - static::getReservedPlacesSubEvent($intId, $database);
         }
 
@@ -909,27 +797,20 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
     {
         $objEvent       = static::findByPk($intId);
         $reservedPlaces = 0;
-        if ($objEvent !== null)
-        {
+        if ($objEvent !== null) {
             $objParentEvent = static::findByPk($objEvent->parentEvent);
-            if ($objParentEvent !== null)
-            {
+            if ($objParentEvent !== null) {
                 $objSubscriber = $database->prepare('SELECT * FROM tl_formdata_details WHERE ff_name=? AND value=?')->executeUncached('eventAlias', $objParentEvent->alias);
-                if ($objSubscriber->numRows > 0)
-                {
-                    while ($objSubscriber->next())
-                    {
-                        foreach (deserialize($objParentEvent->subEventFormFields, true) as $strSubEventFormField)
-                        {
+                if ($objSubscriber->numRows > 0) {
+                    while ($objSubscriber->next()) {
+                        foreach (deserialize($objParentEvent->subEventFormFields, true) as $strSubEventFormField) {
                             $objSubEvents = $database->prepare('SELECT * FROM tl_formdata_details WHERE ff_name=? AND pid=?')->limit(1)->executeUncached(
                                 $strSubEventFormField,
                                 $objSubscriber->pid
                             );
-                            if ($objSubEvents->numRows > 0)
-                            {
+                            if ($objSubEvents->numRows > 0) {
                                 $subEventIds = deserialize($objSubEvents->value, true);
-                                if (in_array($intId, $subEventIds))
-                                {
+                                if (in_array($intId, $subEventIds)) {
                                     $reservedPlaces += 1;
                                 }
                             }
@@ -944,12 +825,9 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
 
     public static function getNonSubEventEvents($blnPublishedOnly = false)
     {
-        if ($blnPublishedOnly)
-        {
+        if ($blnPublishedOnly) {
             return static::findBy(['(parentEvent IS NULL OR parentEvent="")', 'published=1'], []);
-        }
-        else
-        {
+        } else {
             return static::findBy(['(parentEvent IS NULL OR parentEvent="")'], []);
         }
     }
@@ -971,10 +849,8 @@ class CalendarPlusEventsModel extends \CalendarEventsModel
         $objAlias = static::findBy('alias', $varValue);
 
         // Check whether the alias exists
-        if ($objAlias !== null)
-        {
-            if (!$this->id)
-            {
+        if ($objAlias !== null) {
+            if (!$this->id) {
                 return $this;
             }
             $varValue .= '-' . $this->id;